Complete defineProperty synthesis and storage stages
